Add XH_finalCleanUp() hooks
We add the ability to register hooks that will be called at the start
and the end of `XH_finalCleanUp()`, respectively, so that extensions
are able to manipulate the generated HTML (for instance, cleaning it up
with Tidy or HTMLPurifier), or do other tasks related to the generated
HTML.

Inside of `XH_finalCleanUp()` the current working directory might be
reset by the webserver, so we also introduce `XH_CWD` which is supposed
to be used for file system access.

We are not convinced that `XH_beforeFinalCleanUp()` is really useful,
but we implement it for testing and evaluation purposes. Also, we are
not happy with the names (`XH_finalCleanUp()` appears to be to
theoretical), but we stick with it for now. Let's see how this all
turns out.