First stab at optional clean URLs
The basic idea is to move the page name (`$su`) from the query string
into the path.  We let mod_rewrite do the inverse operation.  The
details are already described in the CMSimple Forum for XH 1.6:
https://cmsimpleforum.com/viewtopic.php?f=4&t=5719&start=40#p39249.  We
make the necessary adjustments for XH 1.7.

To actually enable clean URLs, the hidden config setting uri_clean has
to be enabled, and the file `_htaccess` has to be renamed to
`.htaccess`.  The latter obviously requires `.htaccess` support and in
particular mod_rewrite.