Don't include the page URL for the first published page
The first published page of a CMSimple_XH installation can be requested
with or without its page URL (unless we're in admin mode).  To avoid DC
we omit the page URL if possible.  We're doing some trickery in
loginlink(), so that the page URL is always passed there – otherwise
we'd get the unpleasant "No page selected" notice.