Switch from Doxygen to phpDocumentor
phpDocumentor 3.1.2 is still not ready.  The guides somehow work, but
there are still issues, and the feature is marked as experimental.  In
addition, @globals are still not supported and likely never be, so we'd
need to work around.  The default template is not bad, but not really
great either, and none of the other templates support guides.  Also
plantuml rendering appears to be broken.  And the documentation is bad.

TL;DR: we should keep an eye on phpDocumentor, but have to stick with
Doxygen for now. :(